Make server handle multiple clients at once

+++ minitalk/server.c
@@ -29,7 +29,7 @@ void	next_byte(void)
 	g_payload->bit_i = 0;
 	if (!curr_char)
 	{
-		ft_printf("%s\n", g_payload->s);
+		write(1, g_payload->s, ft_strlen(g_payload->s));
 		g_payload->byte_i = 0;
 		ft_bzero(g_payload->s, BUFFER_SIZE);
 		kill(g_payload->sender_pid, SIGUSR2);
@@ -38,7 +38,7 @@ void	next_byte(void)
 	(g_payload->byte_i)++;
 	if (g_payload->byte_i % (BUFFER_SIZE - 1) == 0)
 	{
-		ft_printf("%s", g_payload->s);
+		write(1, g_payload->s, ft_strlen(g_payload->s));
 		ft_bzero(g_payload->s, BUFFER_SIZE);
 		g_payload->byte_i = 0;
 	}
@@ -62,6 +62,8 @@ void	handle_bit(int signo, siginfo_t *info, void *context)
 
 	(void)context;
 	sender_pid = info->si_pid;
+	if (g_payload->sender_pid != sender_pid && g_payload->byte_i != 0)
+		return ;
 	g_payload->sender_pid = sender_pid;
 	process_bit(signo == SIGUSR2);
 	kill(sender_pid, SIGUSR1);
@@ -71,10 +73,12 @@ int	main(void)
 {
 	pid_t				pid;
 	t_stream			s;
+	char				buff[BUFFER_SIZE];
 	struct sigaction	sa;
 
 	ft_bzero(&sa, sizeof(sa));
 	ft_bzero(&s, sizeof(s));
+	ft_bzero(buff, sizeof(buff));
 	g_payload = &s;
 	sa.sa_sigaction = handle_bit;
 	sa.sa_flags = SA_SIGINFO;
@@ -82,7 +86,7 @@ int	main(void)
 	sigaction(SIGUSR2, &sa, NULL);
 	pid = getpid();
 	ft_printf("PID [%d]\n", pid);
-	g_payload->s = ft_calloc(BUFFER_SIZE, sizeof(char));
+	g_payload->s = buff;
 	while (1)
 		pause();
 	return (0);
